relocate lambda
- dropped lambda file into wrong directory
